+++ core/pages/log.php
@@ -38,6 +38,9 @@ class LogPage extends Page
 		// Load lang files
 		$langs->load("log_page@".$dolibase_config['main']['path']);
 
+		// Add CSS files
+		$this->appendToHead('<link rel="stylesheet" type="text/css" href="'.dolibase_buildurl('/core/css/banner.css.php').'">'."\n");
+
 		parent::__construct($page_title, $access_perm);
 	}